pyproject.toml
[tool.poetry]
name = "files-backuper"
version = "0.1.0"
description = ""
authors = ["Moffaty <goga459972@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.166.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pydantic = "^2.11.3"
pydantic-settings = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
